Baud rate and success function
After issuing a command to change the compass baud rate, we need to
change the USART baud rate of the chip to be able to interpret the
messages sent from the compass.

A function to determine if the previous command sent to the compass was
a success, was also added in this commit.